reduce_store: Add a comment regarding a known BUG
I tried to fix it using https://github.com/barkimedes/go-deepcopy, in simple cases
worked but in complex cases (the one I was having the issue for) was not working.

Checking with the original 'flux' https://github.com/facebookarchive/flux/blob/main/src/stores/FluxReduceStore.js#L51
they have a specific comment regarding this, that if it's not immutable you shold not use this.

In Go we cannot have immutability and deepcopy did not work so I added the BUG key
on the documentation.

Changes on this state would affect dicreclty the state and can potentially bring
concurrency issues (this was my issue) as if you have a map init then you could
be writting and reading at the same time and fails.

The solution is then delegating this to the callers, to use Mx internally or helpers
on the Store to return the needed data without returning the satate